// src/app/catalog/models/toolbar/toolbar-file-item.enum.ts
export enum ToolbarFileItemEnum {
  DOWNLOAD_FILE = 'downloadFile',
  DOWNLOAD_XML = 'downloadXml',
  DOWNLOAD_JPEG = 'downloadJpeg',
  DOWNLOAD_SVG = 'downloadSvg',
  SAVE_VERSION = 'saveVersion',
  NEW_FILE = 'newFile',
  GRANT_ACCESS = 'grantAccess',
  COPY = 'copy',
  RENAME = 'rename',
  MOVE = 'move'
}