// src/app/catalog/models/toolbar/toolbar-movement.enum.ts
export enum ToolbarMovementEnum {
  MOVE_UP = 'moveUp',
  MOVE_UP_ACCELERATED = 'moveUpAccelerated',
  MOVE_LEFT = 'moveLeft',
  MOVE_LEFT_ACCELERATED = 'moveLeftAccelerated',
  MOVE_DOWN = 'moveDown',
  MOVE_DOWN_ACCELERATED = 'moveDownAccelerated',
  MOVE_RIGHT = 'moveRight',
  MOVE_RIGHT_ACCELERATED = 'moveRightAccelerated',
}